feat: add standalone DataTable component
